# internal/httpapi/openapi.yaml
openapi: 3.1.0
info:
  title: muni API
  version: 0.1.0
  description: Offline-first collaborative document REST API. Mutating browser requests use an HttpOnly session cookie; integrations use a scoped personal bearer key.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - bearerKey: []
paths:
...
  /auth/login:
    post:
      security: []
      summary: Create a local browser session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identity, password]
              properties:
                identity: { type: string }
                password: { type: string, format: password }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Error' }
...
  /workspaces:
    get:
      summary: List accessible workspaces
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      summary: Create a team workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string, maxLength: 80 }
                slug: { type: string, pattern: '^[a-z0-9][a-z0-9-]{1,46}[a-z0-9]$' }
                description: { type: string }
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /workspaces/{workspaceId}/documents:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      summary: List accessible documents in a workspace
      parameters:
        - { name: trash, in: query, schema: { type: boolean } }
        - { name: folderId, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: string, enum: [DRAFT, REVIEW, PUBLISHED, ARCHIVED] } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /workspaces/{workspaceId}/folders:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      summary: List the workspace folder tree
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      summary: Create a root or nested folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 120 }
                parentId: { type: [string, 'null'], format: uuid }
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /workspaces/{workspaceId}/members:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      summary: List workspace members
      responses:
        '200': { $ref: '#/components/responses/Success' }
    put:
      summary: Add a member or change a non-owner role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, role]
              properties:
                userId: { type: string, format: uuid }
                role: { type: string, enum: [MANAGER, MEMBER, VIEWER] }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /workspaces/{workspaceId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      summary: Remove a non-owner workspace member
      responses:
        '204': { description: Member removed }
  /documents:
    get:
      summary: List recent, favorite, shared, owned, or trashed documents
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [recent, favorites, shared, owned, trash] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      summary: Create a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId]
              properties:
                workspaceId: { type: string, format: uuid }
                folderId: { type: [string, 'null'], format: uuid }
                title: { type: string, maxLength: 240 }
                content: { $ref: '#/components/schemas/TiptapDocument' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /import:
    post:
      summary: Import PDF, DOCX, Markdown, TXT, or HTML as a document
      description: >-
        Headings, lists (including nesting and task lists), tables with merged
        cells, inline formatting, links and embedded images are preserved for
        every format. Images become attachments on the created document.
        Remote image URLs are not fetched. PDFs without a text layer are
        rejected with IMPORT_PARSE_FAILED.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [workspaceId, file]
              properties:
                workspaceId: { type: string, format: uuid }
                folderId: { type: string, format: uuid }
                title: { type: string, maxLength: 240 }
                file:
                  type: string
                  format: binary
                  description: .pdf, .docx, .md, .markdown, .txt, .html or .htm
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
  /documents/{documentId}:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      summary: Read a document after ACL evaluation
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
    put:
      summary: Save document content or metadata using optimistic revision control
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expectedRevision: { type: integer, minimum: 1 }
                title: { type: string, maxLength: 240 }
                content: { $ref: '#/components/schemas/TiptapDocument' }
                status: { type: string, enum: [DRAFT, REVIEW, PUBLISHED, ARCHIVED] }
                visibility: { type: string, enum: [RESTRICTED, WORKSPACE, ORGANIZATION, LINK] }
                reason: { type: string, maxLength: 100 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
    delete:
      summary: Soft-delete a document
      responses:
        '204': { description: Moved to trash }
  /documents/{documentId}/comments:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      summary: List comment threads
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      summary: Add a range-linked comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                parentId: { type: [string, 'null'], format: uuid }
                anchor: { type: [object, 'null'] }
                body: { type: string, minLength: 1, maxLength: 5000 }
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /documents/{documentId}/permissions:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      summary: List explicit document permissions
      responses:
        '200': { $ref: '#/components/responses/Success' }
    put:
      summary: Grant or update a user document permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subjectType, subjectId, role]
              properties:
                subjectType: { const: USER }
                subjectId: { type: string, format: uuid }
                role: { type: string, enum: [EDITOR, COMMENTER, VIEWER] }
                expiresAt: { type: [string, 'null'], format: date-time }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /documents/{documentId}/suggestions:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      summary: List document suggestions
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      summary: Create a range-based change suggestion
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /documents/{documentId}/revisions:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      summary: List immutable document revisions
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /documents/{documentId}/presentations:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      summary: List the presentations made from this document
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
    post:
      summary: Turn the document into a presentation
      description: >-
        The document's structure — headings, lists, tables, figures — is sent to
        the configured presentation service, which designs and builds the deck.
        muni stores only the link between the document revision and the deck;
        the slides and the built file stay with that service.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, maxLength: 200 }
                audience: { type: string }
                purpose: { type: string }
                tone: { type: string }
                language: { type: string }
                slideCount: { type: integer, minimum: 1, maximum: 50 }
                minutes: { type: integer, description: how long the presentation has to run }
                detail: { type: string }
                theme: { type: string }
                templateId: { type: string }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
  /documents/{documentId}/presentations/{presentationId}/status:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - { name: presentationId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Ask the presentation service how generation is going
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/Error' }
  /documents/{documentId}/presentations/{presentationId}/download:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - { name: presentationId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: format, in: query, schema: { type: string, enum: [pptx, pdf], default: pptx } }
    get:
      summary: Download the built presentation
      description: >-
        Streamed from the presentation service on request. muni keeps no copy:
        the deck can be edited there afterwards, and a stored file would go
        stale the moment that happens.
      responses:
        '200':
          description: PowerPoint or PDF file
          content:
            application/octet-stream: { schema: { type: string, format: binary } }
        '404': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
  /documents/{documentId}/presentations/{presentationId}/sync:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - { name: presentationId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: What a document change means for this presentation
      description: >-
        Compares the revision the deck was built from with the current one and
        reports, per slide, whether it should be kept, redrafted, added or
        removed. Nothing is changed; this is what a person reviews first.
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/Error' }
    post:
      summary: Redraft only the slides the document change reached
      description: >-
        Rebuilding a whole deck discards the work someone did in the
        presentation editor. Only the slides whose section changed are put in
        front of the model again; the rest keep their exact text. A conflict
        means the deck was edited in the presentation service meanwhile and the
        request should be retried.
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
  /documents/{documentId}/presentations/{presentationId}/citations:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - { name: presentationId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Write the document and revision each slide came from onto the slide
      description: >-
        A deck that states a number and cannot say where it came from is the
        first thing anyone asks about. muni knows which heading each slide was
        built from, so it writes that as the slide's source line. Cover and
        closing slides make no claim and are skipped, slides that already cite
        something are left alone, and running this twice adds nothing.
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
  /documents/{documentId}/presentations/{presentationId}:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - { name: presentationId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: keepRemote, in: query, schema: { type: boolean, default: false }, description: leave the deck in the presentation service }
    delete:
      summary: Unlink a presentation from the document
      responses:
        '204': { description: Unlinked }
        '404': { $ref: '#/components/responses/Error' }
  /documents/{documentId}/ai/patch:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    post:
      summary: Ask the model to propose block level edits
      description: >-
        The model rewrites whole blocks and the result is stored as pending
        suggestions anchored to the block each one is about. Nothing is applied:
        the proposals join the same review queue a colleague's suggestion does,
        and a person accepts or rejects each one. Only blocks that carry a
        stable id can be proposed against.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instruction]
              properties:
                instruction: { type: string, description: how the document should be changed }
                blockIds:
                  type: array
                  description: limit the proposal to these blocks
                  items: { type: string }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
  /documents/{documentId}/revisions/{from}/diff/{to}:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - { name: from, in: path, required: true, schema: { type: integer, minimum: 1 } }
      - { name: to, in: path, required: true, schema: { type: integer, minimum: 1 } }
    get:
      summary: Compare two revisions block by block
      description: >-
        Blocks are paired by their stable blockId, so inserting a paragraph does
        not make everything below it read as changed. Each block is reported as
        unchanged, added, removed, changed or moved; a changed block carries a
        word-level breakdown.
      responses:
        '200':
          description: Block level comparison
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      from: { $ref: '#/components/schemas/RevisionRef' }
                      to: { $ref: '#/components/schemas/RevisionRef' }
                      summary:
                        type: object
                        properties:
                          added: { type: integer }
                          removed: { type: integer }
                          changed: { type: integer }
                          moved: { type: integer }
                          unchanged: { type: integer }
                      blocks:
                        type: array
                        items: { $ref: '#/components/schemas/BlockDiff' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /documents/{documentId}/export/{format}:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - name: format
        in: path
        required: true
        schema: { type: string, enum: [docx, pdf, html, md, txt] }
    get:
      summary: Export a document
      description: >-
        DOCX exports carry styles, numbering, tables, images and hyperlinks.
        PDF export renders the document through headless Chromium. Both can be
        disabled by administrator policy.
      responses:
        '200':
          description: Exported binary or text document
          content:
            application/octet-stream: { schema: { type: string, format: binary } }
        '403': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /documents/{documentId}/workflow/submit:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    post:
      summary: Submit the current revision when approval workflow is enabled
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
  /search:
    get:
      summary: ACL-filtered PostgreSQL full-text search
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /approvals:
    get:
      summary: List approval requests visible to the current reviewer
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [PENDING, APPROVED, REJECTED] } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
  /admin/ai-usage:
    get:
      summary: AI 호출 기록과 합계
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: muni_session
    bearerKey:
      type: http
      scheme: bearer
      bearerFormat: muni personal API key
  parameters:
    DocumentId: { name: documentId, in: path, required: true, schema: { type: string, format: uuid } }
    WorkspaceId: { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
  schemas:
...
    TiptapDocument:
      type: object
      required: [type]
      properties:
        type: { const: doc }
        content:
          type: array
          items: { type: object, additionalProperties: true }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
  responses:
    Success:
      description: Successful envelope
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: {}
    Error:
      description: Structured API error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }